aps_backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-mptt==0.14.0
drf-yasg==1.21.7
